Rework netbuf_write: * Data is copied into aggregation buffers in netbuf_write_write instead of being queued and only copied immediately before being sent.  (As a result, routines calling netbuf_write_write can release their buffers immediately rather than needing to hold them until the write has been performed.) * Per-write callbacks are no longer performed; instead, a failure callback is provided to netbuf_write_init which gets invoked once when writes fail. * Since there are no per-write callbacks any more, netbuf_write_destroy is no longer needed.
This has considerable fallout in upper layers of code.

In wire_writepacket:
* Since we no longer need to wait for a callback from netbuf_write, the
writdone callback function and struct writepacket are eliminated.
* Since the packet no longer needs to remain valid until a callback occurs,
wire_writepacket no longer needs a callback parameter.

In wire_requestqueue:
* Since we no longer get a "request packet has been written" callback from
wire_writepacket, we can kill writpacket.
* Since there's now only one non-failure code path which leads to us invoking
the "here's the response" callback (when we read a response packet), kill the
cbdone function and merge its functionality into gotpacket.
* Since we're always invoking our callback at the moment that we read the
response packet, we don't need to store the response buffer etc. in struct
request; strip that structure down to the bare essentials of callback+cookie.
* Add a new 'failreq' function which invokes the upstream callback with a
failure status (response buffer = NULL).  Use this in failqueue (instead of
the now-deceased cbdone function).
* Make failqueue take a void * cookie, and provide it to netbuf_write_init as
the writing-requests-failed callback.
* Remove calls to netbuf_write_destroy; it no longer exists.
* Free the buffered writer from failqueue instead of wire_requestqueue_free;
this is necessary to avoid having failqueue invoked twice in the event that
we both fail to read from and fail to write to a connection.
* Check if the request queue is alive before adding a new request to the queue;
we can't just wait for a failure callback from netbuf_write any more.
* In wire_requestqueue_add, construct a packet structure on the stack and pass
it to wire_writepacket, since it no longer needs to be long-lived.

In proto_*_client:
* Move packet buffers from struct foo_cookie onto the stack, since they no
longer need to be long-lived.
* Where packet buffers are variable-length and allocated by malloc, free them
promptly rather than in a callback.

In proto_*_server:
* Remove "packet has been written" callbacks parameters to functions.
* Remove struct packetwrite and packetwrit callback function.
* Move packet buffers onto the stack since they no longer need to be
long-lived.
* Where packet buffers are variable-length and allocated by malloc, free them
promptly rather than in a callback.

In mux:
* Merge contents of callback_writresponse into callback_gotresponse, since we
no longer need to hold onto the response packet until a callback arrives.

In lbs:
* Make dropconnection take a void * instead of struct dispatch_state *, so
that it can be passed to netbuf_write_init as the function to be called if it
becomes impossible to send responses back to the client.
* Kill dispatch_writresponse and don't pass it to proto_lbs_response routines;
instead, decrement the "number of requests pending" value prior to sending
each response back.

In kvlds:
* Remove callback_packet from dispatch_nmr_launch and dispatch_mr_launch, and
don't pass it into proto_kvlds_response_* (as those functions no longer issue
packet-written callbacks).
* Kill writresponse callback function (which was passed as callback_packet),
and instead adjust the requests-in-flight value and possibly read more packets
in callback_mr_done and callback_nmr_done.
* Add new mr_reqs value so we know how much to adjust the requests-in-flight
value in callback_mr_done.
* Make dropconnection take a void * instead of struct dispatch_state *, so
that it can be passed to netbuf_write_init as the function to be called if it
becomes impossible to send responses back to the client.
* Replace 'readreq' function, which unconditionally started a request read, by
a new 'readreqs' function which starts a request read if there isn't already
a request read in progress, there aren't MAXREQS requests in flight, and the
connection isn't dying.
* Replace instances of "call readreq if we're not at MAXREQS requests" with
unconditional calls to readreqs.
* Remove 'accepting' flag; add a 'dying' flag which is set by dropconnection
(if we fail to read a request or the buffered writer for responses fails).
* Simplify dispatch_alive predicate to "we're not trying to die or we have
requests left to handle before we can die".

In total the write buffering changes and elimination of write-completion
callbacks provides a speedup of about 5-15% while reducing the code size by
about 2%.