Complain about HTTP 200 response to RANGE request
While it is technically valid to send an HTTP 200 response to a GET
request which contains a Range: header, it is unexpected (since it
wastes bandwidth) and not supported by kivaloo's Amazon S3 code at
the present time.

This may change in the future; but for now print a warning so that
it's clear why we're treating the request as failed.