Fix 0-byte memory leak when no leaves are dirtied
If none of the MRs in a batch take effect (e.g., there's just one ADD,
and the key already exists) then no pages would be dirtied and we
would set B->dirties to malloc(0).  If the cleaner also opts to not
dirty any pages, then callback_gotleaves would skip batch_run -- and
as a result, B->dirties would never be freed.

On platforms where malloc(0) returns NULL ("a silly answer to a silly
question"), this was fine; but when malloc opts for the alternative
allowed by POSIX (return a unique non-NULL pointer) it means that we
were leaking the (zero byte!) allocation.

We now check if batch_dirty dirtied any pages, and skip the allocation
if not (but set the pointer to NULL, since it might get freed anyway,
depending on the cleaner behaviour).