Detect compilers which lack the restrict keyword.
For now all we do is check if -std=c99 fixes the problem and add
that flag.  (We don't want that option by default since -std=c99
is not a POSIX option.)  At some point in the future we might
want to add a -Drestrict= flag if -std=c99 doesn't convince the
compiler to behave in a C99-compliant way.