libcperciva: allow ${CFLAGS_POSIX} with out-of-tree builds
The previous version of posix-cflags.sh failed if we did not have write access
to libcperciva/POSIX/.  In general, this would not be a problem, but autotools
is designed to allow "out-of-tree" builds, where the compilation happens in a
separate directory.  It seemed like a shame to break our out-of-tree-capable
status when it could be fixed so simply.