Switch s3_request() to AWSv4 signing.
The older signing scheme is not available in new regions.

The downside to this change is that AWSv4 has per-region signing
keys (derived from a master key) and so s3_request() must now be
aware of which AWS region the bucket it's accessing is located in.

This will have cascading effects through users of this code.