Add routines for using DynamoDB as a KV store.
To be precise, using string keys in a field named "K" and binary values
in a field named "V".

This can be used directly, but we will be using it with "values" which
are B+Tree pages, since DynamoDB's I/O pricing is constant for items
up to 1 kB in size -- this allows us to pack multiple key/value pairs
into a single DynamoDB operation.