It's ok if pages !exist when scanning for a root
Add btree_node_fetch_try, which is like btree_node_fetch but
treats an LBS "page does not exist" response as a non-error.

Use this during B+Tree initialization when we're scanning for
a root node; this makes it ok if LBS crashes after writing an
arbitrary subset of pages in an APPEND request, as long as the
final (root) page is not written until the rest of the pages
are safely stored.