Rework DynamoDB request queue rate limiting
Switch from targetting DynamoDB requests per second to targetting
DynamoDB capacity units per second.

1. Extract ConsumedCapacity->ConsumedCapacity from DynamoDB responses,
add it to the request log, and keep a rolling average of the average
number of capacity units used per request.

2. Track an estimated 'burst capacity available' by
(a) Zeroing the estimate when we get a ProvisionedThroughputExceeded,
(b) Incrementing based on the provisioned capacity, and
(c) Decreasing based on reported capacity used; and
use this (and the rolling average of capacity units per request) to
limit the number of concurrent requests.

3. When running the request queue, keep sending requests as long as
we're below our burst limit, rather than sending at most one request
at once.

4. Remove the "ops/s" parameter from dynamodb_request_queue_init.
Queues default to unlimited throughput (as with "on-demand" DynamoDB
tables).

5. Add a new "setcapacity" function; this can be called at any time,
in case the provisioned capacity of the table changes.