Monitor DynamoDB table provisioned capacity
Issue DescribeTable requests to determine whether a DynamoDB table
is configured for on-demand or provisioned capacity; if it's set for
provisioned capacity, pass those values to the DynamoDB request queues
so that they can rate-limit requests appropriately.

This commit also catches up to API changes in 4cfb5773 -- the
dynamodb_request_queue_init function no longer takes a capacity value
when the queue is initialized, but instead there is a setcapacty
function which is called later.  (Precisely so that we could use it
right now.)