Add timeouts to DynamoDB request queue.
Right now the timeouts are fixed at 1/2/4/8/15/15/15/... seconds; a
future commit will add tracking of request time statistics so that
the timeouts can be adjusted automatically at run time.

We also use the timeouts for the purpose of back-off in the event
of request failures; if a request fails (and not with an error which
we want to pass upstream) then we wait until the timeout expires
before sending a new request.  This is particularly important in the
event of network errors -- otherwise we can try to send a request
thousands of times per second (and fill our log files with records
of requests which failed after a handful of microseconds).

Non-trivial consequences of these changes:
* A request is now sendable if it has no *timeout* cookie, regardless
of whether it has an *http* cookie.
* The "HTTP request done" code path has been churned a bit in order
to avoid unnecessary code duplication with the (newly added) "timer
expired" code path.