Compute dynamodb timeouts from past statistics
We closely follow the approach used for TCP timeouts, with the
exception that where TCP uses [mean] + 4*[mean deviation], we use
1.5*[mean] + 4*[mean deviation].  This is necessary since we're
running over TCP; while the 4*[mean deviation] term is sufficient
to accommodate queuing delays, any lost packets in a TCP connection
will add a minimum of 1 RTT to the request.  Since DynamoDB requests
and responses often fit into a single TCP segment, there may be as
little as 2 RTT before we have the response:
	SYN	-->
		<-- SYN/ACK
	request -->
		<-- response
	--- at this point the client forgets about the connection ---
	--- and moves on with its work, leaving the kernel to     ---
	--- clean up the TCP connection.                          ---
	FIN	-->
		<-- FIN/ACK
	ACK	-->

As a result, in order to accommodate the possibility of a packet being
dropped, we need at least an extra 1 [network RTT] / 2 [network RTT]
i.e., 50% of the request RTT, on top of TCP's 4*md accommodation for
packet queuing delays.  (If we ever start using TCP Fast Open, this
would need to be increased to 100% since that could yield responses
after 1 network RTT.)

Note that this does not account for the time taken by the DynamoDB
service between receiving our request and sending the response back;
from userland there is no way to distinguish this part of the overall
request round trip from the amount which is strictly *network* RTT.
(In the kernel it is possible to distinguish, at least given the
current configuration of DynamoDB hosts, since they send a TCP ACK
for the segment contaning the request immediately rather than waiting
for the response to be ready.  This is arguably a misconfiguration in
DynamoDB, since it wastes bandwidth and CPU time.)