Correctly assert lengths read by serverpool
This was a = / == typo; we want to assert that readlen is correct,
not set it to the right length.

Fortunately this was mostly harmless: It happens when asserting
the length we're reading over a socket from ourselves, and readlen
would only have its value changed if the assertion was about to
fail.

Reported by:	Coverity Scan