Update 'nextblk' before appending, not after
Update the 'nextblk' field in the lbs-dynamodb state at the beginning
of the append operation (to be precise, before storing the new value)
rather than storing the new value but only updating it in our internal
state after the append completes.

Note that there is a slight change in behaviour: If a PARAMS request
is made while an APPEND request is still in progress, it now receives
the *new* "next block" and "last block" values rather than the *old*
values.  This reduces, but does not eliminate, a race condition where
if KVLDS crashes and is replaced it may issue a request which is
rejected as invalid -- formerly because it attempts to append at the
wrong position, but now only if it attempts to append before the
previous append completes.

While we're here, rework some of the documentation related to how the
"next block" and "deleted to" values work.