Refactor metadata handling in lbs-dynamodb
The metadata_* code now holds a pointer to the dynamodb-kv request
queue.  When the deleteto and state code is initialized, it is
passed a pointer to the metadata state, and that is used to make
metadata read/write requests.

The request dispatcher now calls directly into deleteto to handle
FREE requests; consequently, state_gc is no longer needed and the
state_* code no longer needs a pointer to the deleteto state.

No functional changes.