Merge nextblk and deletedto into a single item
Keep a single "metadata" DynamoDB item which holds both the
"nextblk" and "deletedto" values.

When "nextblk" needs to be updated, store an updated metadata
as soon as we're not already storing one.

When "deletedto" needs to be updated, wait up to 1 second before
kicking off the write, since it's (a) not in the blocking path
for anything urgent, and (b) can often be combined with storing
an updated "nextblk" value.