Add "wear leveling" for lbs-dynamodb metadata
Store lbs-dynamodb metadata in a series of 4096 items, writing updated
values into items in turn; on startup, use a binary search to find the
most recently stored values.  This prevents us from having a "hot spot"
and associated imbalance in per-partition DynamoDB loads due to the
repeated rewriting of the metadata.