Explicit (size_t)random() and (size_t)rand()
These two functions return a value between 0..RAND_MAX, but stored as a
signed integer.  We're always taking that value mod (size_t), which
mixes signed and unsigned integers.

This code assumes that RAND_MAX <= SIZE_MAX, which I think is safe
enough that we don't need a CTASSERT() since this is for
non-cryptographic stuff.