Adjust DynamoDB capacity-increase timer condition
Prior to this commit we kept the timer running whenever the
bucket wasn't overflowing; we now start the timer only when
the bucket is empty.  This change is made possible because
we're no longer relying on the timer constantly ticking; when
it ticks we "catch up" with however much capacity will have
accumulated.