Unlock a unclean nodes the cleaner reaches
The cleaner periodically finds the oldest node and examines it
and its siblings to see if they should be rewritten during the
next batch.  If this set happens to include a btree node which
is already being dirtied, the cleaner can ignore that node --
it will be rewritten anyway.  In that case, the cleaner needs
to drop its lock on the node.

Prior to this commit, the above scenario results in an assertion
failure due to a shadow page being freed (after the batch is
committed) with a lock held (by the cleaner).