Limit memory used for pages being cleaned
If we're using more than 1/16 of our memory to hold pages which
are being cleaned, don't look for any more pages to clean; if we
try to clean too many pages at once, we could end up locking too
many pages in memory.

This is primarily a theoretical concern: In order to come close
to this limit we would need to have a very small page pool, and/or
be cleaning very aggressively and have extremely slow writes --
otherwise there's no way the cleaner can accumulate enough pages
before they are released as part of a batch commit.