Adjust handling of "no old leaves" case
Rather than merely checking for the case where all the leaves under a
height-1 node are already being touched, check to see if the oldest
leaf is old enough that we would clean it.

This allows us to convert an error -- which should never have been an
error since it was harmless -- into an assertion that we find at least
one cleanable page.