Don't re-find new leaves repeatedly
If the oldest leaves are too new to be cleaned -- which can only
happen if half of the btree is already being cleaned, i.e., we
have a small tree and/or very aggressive cleaning -- then there's
no point poking the cleaner to look for more leaves to clean; it
will just find the same (ineligible) leaves repeatedly.

In the "no leaves are eligible for cleaning" case, skip the call
to poke the cleaner to do more work.