lbs: explicit cast to (off_t)
We know that
	blkno - fs->start <= fs->len
and fs->len is the length of the file in blocks.  Assuming that the
maximum length of a file is fits into off_t [1], that means that
	fs->len * S->blocklen <= OFF_MAX
and therefore
	(blockno - fs->start) * S->blocklen <= OFF_MAX

[1] theoretically we might have been in trouble if somebody created lbs
on a 64-bit platform (which also defined the non-standard OFF_MAX [2]),
then copied the files over to a 32-bit filesystem platform.  However, if
an overly large file is read in storage_findfiles(), the lstat() call
will fail with EOVERFLOW.  So if we reach this part of the code, we know
that the file length fits into off_t.

[2] if OFF_MAX is not defined (as is common on Linux), storage_init()
assumes that files cannot be larger than INT32_MAX.