Add network_ssl subsystem
This is a wrapper around non-blocking SSL library code which matches the
API of network_read/network_write in libcperciva.

Only the client side is implemented here; servers should unwrap their
connections using hitch or a similar tool.

The root certificates can be specified using the environment variables
    SSL_CERT_DIR
    SSL_CERT_FILE
If neither of these are set, the default ones for OpenSSL will be used.

A minimum protocol version of TLS 1.2 is used.

SSL stack errors are reported to stderr.

OpenSSL 1.1.1 or later is required; compatibility with earlier versions
and/or other SSL stacks may be added later.