REBASE keep return value from SSL_write_ex, SSL_read_ex
It's true that in the man-page for openssl 1.1.1, it says that these
functions return 1 for success and 0 for failure.  So the original code
isn't wrong.

However, earlier versions distinguish between 0 and -1 failure cases,
and the docs for SSL_get_error() clearly states that it wants the return
code.  Supporting this is vital if we want to allow openssl < 1.1.1.

(also, who knows how these functions may change in the future?  seems
safer to give SSL_get_error() exactly what it wants: the return code.)