lib/proto_s3: S3 GET responses can include an error status
Looking at s3/dispatch.c, callback_response(), case PROTO_S3_GET (line 268):
if we don't have a body, or have an invalid ETag, the status is set to 0
and we call proto_s3_response_get().

That function itself is #define'd as proto_s3_response_data(), prompting
the comment change here.  The phrase "(or 0 on error)" was used for
proto_s3_request_head().