Use (long) for Dynamodb _setcapacity()
The code which parses ReadCapacityUnits and WriteCapacityUnits uses a
(long), but that's truncated to an (int) when they're passed to
dynamodb_request_queue_setcapacity().  Internally, that function sets
Q->cappers which is a (double), so there's no harm changing the argument.

It's unlikely that we'll want to do more than INT_MAX operations per
second, of course.