dynamodb-kv: use PARSENUM_EX for capacity
json_find() returns a pointer to an existing buffer, so there will be
other characters after the relevant value.  This function already set
the last byte in this buffer to NUL, so it's safe to use PARSENUM_EX.

A failure to parse the value does not indicate an internal error, so we
jump to doneparse rather than err0.

The AWS API says that these values have a minimum value of 0:
https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_ProvisionedThroughputDescription.html#DDB-Type-ProvisionedThroughputDescription-ReadCapacityUnits