http: use PARSENUM_EX for chunked headers
The cast to (unsigned char *) is slightly dangerous because buf is not
NUL-terminated.  However, we've already checked that the buffer contains
an EOL, and we know that PARSENUM_EX will stop at the first character
which is not in base 16.  So the cast is safe.

Full disclosure: I added that cast in 69b70b418b790f1a4438359d41b1366956bc88d1
without realizing the danger.