build: avoid empty top-level -config.sh files
I'm sometimes confused when I try compiling on a raspberry pi because
there's a 0-byte cpusupport-config.h.  We don't support any special ARM
features, so that makes sense... but my first thought is always that
something went wrong in the build system.

We'll only ever end up with a blank posix-flags.sh if we don't include
libcperciva/POSIX, and I doubt that we'll ever have a project where we
use this build system but don't want to check POSIX compatibility.  But
I figured that we might as well use the same boilerplate for both config
files.