Handle responses without ConsumedCapacity.
DynamoDB's behaviour with "ReturnConsumedCapacity": "TOTAL" is rather odd:

* A PutItem which returns a ConditionalCheckFailedException error consumes
write capacity but is documented as not returning a ConsumedCapacity field.

* A request which returns a ProvisionedThroughputExceededException used to
return a ConsumedCapacity of zero, but at some point in 2019 or 2020 this
(without any corresponding change in documentation) changed to omitting the
not returning ConsumedCapacity field.

We treat a lack of ConsumedCapacity as zero; for a failed conditional
operation this will not be correct, but it's the best we can do under the
circumstances (and hopefully failed conditional operations won't occur very
often).