Disallow PARAMS/PARAMS2 while APPEND in progress
Allowing this could only lead to problems:
* If the returned "last block stored" value reflects the data currently
being written, the caller might attempt to read blocks which have not
in fact been stored (and might never be stored, in the event of an I/O
error).
* If the returned "last block stored" value does not reflect the data
currently being written, that data is effectively orphaned -- we would
not be able to return success from the APPEND operation since we have
indicated that the data in question does not exist.

Since lbs completes all operations after a connection closes before it
accepts a new connection, and kvlds only makes a PARAMS request when
it first connects, this commit should have no effect in the normal
lbs+kvlds setup; however, it could prevent problems with lbs+mux+kvlds.