lbs-dynamodb: Store metadata in a separate table
In 7e12a018 we changed how metadata was stored to "wear level" it
across many items; this made it possible to avoid the DynamoDB "hot
item" problem, since no item was accessed often enough to have any
risk of unbalancing the DynamoDB partitions.

This commit reverts that, aside from the addition of the "generation"
value (which will be independently useful), and adopts a much simpler
solution: The metadata item gets stored in a different DynamoDB
table.  As a consequence, lbs-dynamodb now takes *two* dynamodb-kv
daemons as parameters -- one for the data blocks and one for the
metadata item.