lbs-dynamodb: Track 3 versions of metadata
We now have "stored" metadata (what's currently known to be stored in
DynamoDB), "storing" metadata (what we're currently in the process of
storing in DynamoDB), and "latest" metadata (values we've been told
about).  The _write functions update "latest" values, while the _read
functions return the "stored" values.

This is a bookkeeping change only: We should not currenly be reading
values at a time when the distinction between stored/storing/latest
matters.