lbs-dynamodb: Rework deleteto system
Rather than issuing deletes for a batch of 256 items, waiting for those
to complete, then issuing a metadata update and waiting for that to
complete, we now update the metadata asynchronously with respect to the
deletion of items.  Deletes will be issued whenever
* There are blocks which aren't needed,
* We have fewer than 64 deletes currently in progress, and
* Fewer than 8000 deletes are "unrecorded" in the sense of the stored
metadata not reflecting that they have been performed.

The value 8000 is chosen so that in the worst case after a crash we spend
$0.01 re-issuing deletes for previously-deleted DynamoDB items.  With
metadata updates happening at least once a second, this allows for 8000
deletes/second ($21k/year of WCUs at current reserved pricing, not counting
the original cost of creating the DynamoDB items), which seems like a
reasonable default.  (If metadata updates happen more often, this limit
will increase.)

The value 64 is chosen to allow the same 8000/second rate if each request
takes 8 ms to complete.

While I'm here, update an API comment: deleteto_init can no longer call
events_run.  (Note that deleteto_stop still can, however, since it waits
for deletes in progress to complete and for updated metadata to be stored.)