lbs-dynamodb: Rework APPEND & track "lastblk"
Prior to this commit, an APPEND operation was "committed" by the
act of writing the last (highest-numbered) block.  This leads to two
issues:

1. If lbs-dynamodb crashed, after restarting kvlds would need to
scan backwards looking for a root page (which is always the highest
numbered block out of successfully completed APPENDs).  This could
take a while on a system which performs large batches of operations.

2. When lbs-dynamodb and kvlds are restarted after a crash, there's
potentially a race if the final block write is in progress (which
includes if it is buffered in dynamodb-kv waiting for provisioned
write capacity quota).

After this commit, an APPEND occurs in three steps:
1. Record the new "nextblk" value, effectively reserving space;
2. Write the blocks;
3. Record the new "lastblk" value, effectively committing the atomic
append operation.

Finding kvlds' root page is now trivial since lastblk always points
to it; and the only relevant race is on the "metadata" item (which will
be addressed in future changes).