kvlds: Don't automatically disable cleaning
Prior to this commit, if lastblk != nextblk - 1, we assumed that the
block ID space was sparse and disabled cleaning since our normal
process for cleaning would not work.  This was necessary when using
lbs-s3 as a page storage backend.

We now have lbs-dynamodb, which if an APPEND is interrupted, can have
lastblk pointing at the last page of the last *completed* APPEND while
nextblk points after the *interrupted* APPEND; this is perfectly
harmless and does not indicate a sparsity to block IDs.

Users of kvlds with lbs-s3 should manually disable cleaning.