kvlds: free findleaf cookie outside of function
In the previous code, findleaf() would free the cookie in some code
paths, but not all:

- NOT FREE, return 0: if the node is not present
- NOT FREE, return -1: if there was an error in findleaf() itself
- FREE, return 0: if the node is present and there is no callback or
  the callback succeeds
- FREE, return non-zero: if the node is present, there is a callback,
  and the callback fails

This made it hard to reason about whether the caller function should
free the cookie.  As a result, in one edge case (node present, callback
function present and returned non-zero), there was a double-free.

This commit removes any mpool_findlead_free() from findleaf(); instead,
we put it in the calling functions of btree_find_leaf() and
btree_find_range().

Reported by:	clang scan-build