lbs-dynamodb: Use atomic ops on metadata item
Use ICAS or CREATE when initially claiming ownership of metadata,
and use ICAS when storing updated metadata.

Handle precondition failures when storing updates by exiting -- this
means that we are being replaced by a new lbs-dynamodb process.

Handle precondition failures when initially claiming ownership by
looping (potentially indefinitely, if we keep on losing races) --
this indicates that we launched the new lbs-dynamodb process in an
attempt to replace a previous one and we want to "steal" ownership
of the metadata.