lib/dynamodb_request_queue: Handle DynamoDB errs
When receiving an HTTP 400 response from DynamoDB, extract the error
string.  Use this string to check if our requests are being throttled,
replacing previous hacky ProvisionedThroughputExceededException
identification code; and pass the error string upstream to users of
the DynamoDB request queuing API.

Currently the users of this API don't do anything with said error
string; that will change in a future commit.