lbs-dynamodb: Clean up metadata write triggering
Add a new metadata_flush function which deleteto code uses to request
a metadata write if the stored "deletedto" value falls too far behind.

The parameters now allow for deletes to be issued at a steady-state
rate of MAX(2000 / T_1, 64 / T_2) where T_1 is the time taken to write
metadata and 64 / T_2 is the time taken to delete an item.  In practice
the second limit (no more than 64 concurrent deletes) is very likely to
be the one which is relevant.

Remove the timer from metadata; we no longer trigger metadata writes
1 second after a new "deletedto" value is recorded.