sha256_sse2: fix API comment for SPAN_ONE_THREE
The code is functionally ok, but the macro's comment had ${a} and ${b}
backwards.  For example, consider the line:

    Xj_minus_fifteen = SPAN_ONE_THREE(X0, X1);

What we want to get is:
    Xj_minus_fifteen := W[j-12] W[j-13] W[j-14] W[j-15]

Our memory layout (and underlying the words we want) is:

    X1:    W[j-9]  W[j-10] W[j-11] W[j-12]
                                   ~~~~~~~
    X0:    W[j-13] W[j-14] W[j-15] W[j-16]
           ~~~~~~~~~~~~~~~~~~~~~~~

So we want three elements of X0 (which is ${a}), and one element of X1
(which is ${b}).