Add lbs-dynamodb-init utility
This creates a pair of DynamoDB tables ("data" and "metadata") for use
with the lbs-dynamodb daemon, and stores an initial metadata Item in
the metadata table.

Note: The two tables are created as on-demand DynamoDB tables; for any
significant load it may be preferable to switch the data table into
"provisioned" mode (most likely with autoscaling rules in place).