lbs-dynamodb: Add tableid to data + metadata
In lbs-dynamodb-init, a 256-bit random value is generated and stored in
the data table with the key "tableid", and as part of the initial metadata
value in the metadata table.

In lbs-dynamodb, the table ID from the metadata is compared against the
value from the data table.

This serves no functional purpose, but merely acts as an anti-foot-shooting
safeguard.