cpusupport: rename CRC32_64 to SSE42_64, sse42
The other cpusupport files are aimed at a specific set of intrinsics
(e.g. SSE2, AENSI), rather than one particular type of function (CRC32).
This change will make it more clear what's happening, and fixes my mistakes in
    2018-08-29 cpusupport: detect CRC32 CPU extensions
    512eb7fdf613f7209a439832eee7bfc460a47279
and
    2019-10-29 cpusupport: rename X86_CRC32 to X86_CRC32_64
    5d76d9e3af3e741afc8364aa63984660a5b344bc

Note that _mm_crc32_u64() is hidden within #ifdef __x86_64__ in
<smmintrin.h> for both clang and gcc.  There are two other intrinsics
which receive the same treatment: _mm_insert_epi64 and _mm_extract_epi64.

The runtime detection of SSE4.2 does not depend on whether we are
compiling 64-bit code or not, so I removed the _64 suffix in the renaming.