cpusupport/Build: detect SSE42 (incl. 32-bit)
cpusupport-X86-SSE42-64.c only detects SSE42 if we're compiling a 64-bit
binary.  That's deliberate, as we can only use _mm_crc32_u64() if that's
the case.

However, in some cases we might want to know if we have SSE42 at all,
regardless of 32-bit or 64-bit; hence the new file.