network_ssl_compat: revise OpenSSL compatibility
There's two goals here:

1) List all defines explicitly in the same place.  This is done by
   having a single #if statement (with multiple #elif), instead of
   multiple #if statements.

   The old code had "if openssl before 1.1.0, use these; if before 1.1.1,
   use this as well".  That's wasn't the best option for readability.
   In the new code, once you find the block corresponding to the library
   version, that's it.

2) Structure the #ifs so that it goes from most recent to earliest.

   I find it more natural to see
       #if version >= 101000
       /* Do stuff required for 1.1.0 */
       ...

       #else
       ...

   rather than
       #if version < 101010
       /* Do stuff required for 1.1.0 */
       ...

       #else
       ...

    where the different version numbers look like a mistake.

As an additional nitpick: the final nibble of the version numbers from 0 to f.
That nibble indicates the "status", namely 1 to e for betas, and f for
release.

Technically, this could break compilation for somebody using a beta
release of an old version of OpenSSL.  But a) it would be insane for
somebody to be doing that, and b) this brings us into line with the
expected value of OPENSSL_VERSION_NUMBER values.  Particularly, LibreSSL
follows the same pattern of ending their LIBRESSL_VERSION_NUMBER values
with an f.