network_ssl_compat: reorganize SSL library #if structure
The previous structure was
    #if LibreSSL
    pretend to be OpenSSL version x.y.
    #endif

    #if OpenSSL version a.b
    ...
    #elif OpenSSL version c.d
    ...
    #endif

Unfortunately that didn't actually work; the LibreSSL API does not
(completely) line up with any OpenSSL version.

The new structure is
    #if LibreSSL
    ... handle LibreSSL versions

    #else
    ... handle OpenSSL versions

    #endif

(As of this commit, there's no "handle LibreSSL versions" code; this is
about setting up the overall structure.  Specific LibreSSL version
support is in the following commit.)