warnp: preserve errno in warn() and warnx()
fprintf(3) and vsnprintf(3) are not documented as being unable to modify
errno during a successful operation, so we must save & restore errno
ourselves in warn() and warnx().

warnx():

(1) if DEBUG is enabled, the warnline; can change errno, either from 0
to non-zero causing the wrong branch in warnp() to be taken, or from the
errno that we should report to some unrelated one.

(2) even when DEBUG is not enabled, the fact that warnx() may set errno
means that the warnp() macro doesn't actually guarantee that errno is 0
afterwards.

warn():

We don't actually need to restore errno here, but it's a good idea in
terms of consistency.

Reported by:	Rasmus Villemoes
Bug bounty:	$20 (alphatest)