mpool.h: compile-time assert for a non-zero size
Previously, we could write
    MPOOL(foo, int, 0);
which would expand to
    static void * mpool_foo_static[0];

An array declaration with fixed size 0 is not allowed in C99:
    6.7.5.2 Array declarators
    ...
    If the expression is a constant expression, it shall have a value
    greater than zero.

gcc complains about such an array, but clang doesn't.

Even if we ignore the zero-size array, the above would also expand to:
    static struct mpool mpool_foo_rec = {0, 0, ...

which would set M->allocsize = 0.  That would result in malloc(0) being
called inside mpool_free(); also, mpool_free() would be unable to
increase the size of the pool since that code relies on doubling
M->allocsize.