proto_lbs_request_append_blks(): forbid nblks==0
According to INTERFACES, an lbs APPEND request must have "1 or more
blocks of data", so a request with 0 blocks would be invalid.

If we call this function with (nblks == 0), then it sends a badly-formed
request to the lbs server, which responds by dropping the connection
(which is its usual response to an invalid request).