{api,cpu}support.sh: use CPPFLAGS as well as CFLAGS
Back in:
    2022-08-30 build: put CPUSUPPORT & APISUPPORT flags after CFLAGS
    dfaebf620429fc612e3300f9512b920efe021053
we ensured that the CPUSUPPORT & APISUPPORT flags would override the
user-specified CFLAGS, because those are required for compilation.

That solution works great for our POSIX Makefiles, but it fails with
autotools.  In particular, automake always puts $(CFLAGS) after all
build system flags:
    https://www.gnu.org/software/automake/manual/html_node/Flag-Variables-Ordering.html

As a result, if we put our "nitpicky enable (almost) all warnings" in
CFLAGS, there's no way for cpusupport-config.h to override them.

The cleanest solution is:
- pass our "nitpicky warnings" in CPPFLAGS
- read from CPPFLAGS (as well) in {api,cpu}support.sh